organized demo project
