+++ lib/screens/home_screen.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+import 'package:numbers_game_demo/screens/game_screen.dart';
+
+class HomeScreen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    final darkBlue = Color.fromARGB(255, 18, 32, 47);
+    return MaterialApp(
+      theme: ThemeData.dark().copyWith(scaffoldBackgroundColor: darkBlue),
+      debugShowCheckedModeBanner: false,
+      home: GameScreen(),
+    );
+  }
+}
