+++ lib/utils/constants.dart
@@ -1,0 +1,5 @@
+import 'package:flutter/material.dart';
+
+const targetColors = [Colors.orange, Colors.green, Colors.yellow, Colors.blue];
+const textColors = [Colors.blue, Colors.yellow, Colors.green, Colors.orange];
+const colorNames = ['orange', 'green', 'yellow', 'blue'];
