+++ lib/utils/target_type.dart
@@ -1,0 +1,1 @@
+enum TargetType { color, number }
